Retype tictoc related variable from float to double in header.

+++ mbircone/src/MBIRModularUtilities3D.h
@@ -358,9 +358,9 @@ struct ParallelAux
 struct SpeedAuxICD
 {
     long int numberUpdatedVoxels;
-    float tic;
-    float toc;
-    float voxelsPerSecond;    
+    double tic;
+    double toc;
+    double voxelsPerSecond;
 };
 
 struct IterationStatistics
